+++ src/google_music_utils/constants.py
@@ -3,15 +3,26 @@ __all__ = ['CHARACTER_REPLACEMENTS', 'FIELD_MAP', 'TEMPLATE_PATTERNS']
 from multidict import MultiDict
 
 CHARACTER_REPLACEMENTS = {
-	'\\': '-', '/': ',', ':': '-', '*': 'x', '<': '[',
-	'>': ']', '|': '!', '?': '', '"': "''"
+	'\\': '-',
+	'/': ',',
+	':': '-',
+	'*': 'x',
+	'<': '[',
+	'>': ']',
+	'|': '!',
+	'?': '',
+	'"': "''"
 }
 """dict: Mapping of invalid filepath characters with appropriate replacements."""
 
 _FIELD_MAP_GROUPS = [
-	('albumartist', 'album_artist', 'albumArtist'), ('bpm', 'beatsPerMinute'), ('date', 'year'),
-	('discnumber', 'disc_number', 'discNumber'), ('disctotal', 'total_disc_count', 'totalDiscCount'),
-	('tracknumber', 'track_number', 'trackNumber'), ('tracktotal', 'total_track_count', 'totalTrackCount')
+	('albumartist', 'album_artist', 'albumArtist'),
+	('bpm', 'beatsPerMinute'),
+	('date', 'year'),
+	('discnumber', 'disc_number', 'discNumber'),
+	('disctotal', 'total_disc_count', 'totalDiscCount'),
+	('tracknumber', 'track_number', 'trackNumber'),
+	('tracktotal', 'total_track_count', 'totalTrackCount')
 ]
 _FIELD_MAP = [(field, alias) for group in _FIELD_MAP_GROUPS for field in group for alias in group if field != alias]
 
@@ -19,10 +30,20 @@ FIELD_MAP = MultiDict(_FIELD_MAP)
 """MultiDict: Mapping of field name aliases."""
 
 # TODO: Support other/more metadata field names.
-# TODO: Support multiple field names per template pattern.
 TEMPLATE_PATTERNS = {
-	'%artist%': 'artist', '%title%': 'title', '%tracknumber%': 'tracknumber',
-	'%album%': 'album', '%date%': 'date', '%genre%': 'genre',
-	'%albumartist%': 'albumartist', '%discnumber%': 'discnumber'
+	'%album%': ['album'],
+	'%albumartist%': ['albumartist', 'album_artist', 'albumArtist'],
+	'%artist%': ['artist'],
+	'%date%': ['date'],
+	'%disc%': ['discnumber', 'disc_number', 'discNumber'],
+	'%disc2%': ['discnumber', 'disc_number', 'discNumber'],
+	'%discnumber%': ['discnumber', 'disc_number', 'discNumber'],
+	'%discnumber2%': ['discnumber', 'disc_number', 'discNumber'],
+	'%genre%': ['genre'],
+	'%title%': ['title'],
+	'%track%': ['tracknumber', 'track_number', 'trackNumber'],
+	'%track2%': ['tracknumber', 'track_number', 'trackNumber'],
+	'%tracknumber%': ['tracknumber', 'track_number', 'trackNumber'],
+	'%tracknumber2%': ['tracknumber', 'track_number', 'trackNumber']
 }
-"""dict: Mapping of template patterns to their audio-metadata field name."""
+"""dict: Mapping of template patterns to their metadata field names."""