+++ CHANGELOG.md
@@ -6,7 +6,12 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased](https://github.com/thebigmunch/google-music-utils/tree/master)
 
-[Commits](https://github.com/thebigmunch/google-music-utils/compare/2.0.1...master)
+[Commits](https://github.com/thebigmunch/google-music-utils/compare/2.1.0...master)
+
+
+## [2.1.0](https://github.com/thebigmunch/google-music-utils/releases/tag/2.1.0) (2019-02-01)
+
+[Commits](https://github.com/thebigmunch/google-music-utils/compare/2.0.1...2.1.0)
 
 ### Added